+++ lib/freeze_the_lits/auto.rb
@@ -1,3 +1,0 @@
-require "freeze_the_lits"
-
-FreezeTheLits.setup(watch_dirs: [Dir.pwd])