+++ lib/freezolite/auto.rb
@@ -1,0 +1,3 @@
+require "freezolite"
+
+Freezolite.setup(watch_dirs: [Dir.pwd])