+++ test/freeze_the_lits_test.rb
@@ -1,7 +1,7 @@
 require "test_helper"
 
 class FreezeTheListTest < Minitest::Test
-  if ENV["FREEZE_THE_LITS"] == "false"
+  if ENV["FREEZOLITE_DISABLED"] == "true"
     def test_fixtures_default_behaviour
       load File.join(__dir__, "./fixtures/app/name.rb")
 
@@ -18,7 +18,7 @@ class FreezeTheListTest < Minitest::Test
       assert_equal "name 1 2", NameTest.run
     end
   else
-    def test_fixtures_with_freeze_the_lits_enabled
+    def test_fixtures_with_freezolite_enabled
       load File.join(__dir__, "./fixtures/app/name.rb")
 
       assert_raises(FrozenError) { Name.new + "test" }
@@ -29,7 +29,7 @@ class FreezeTheListTest < Minitest::Test
       assert result.end_with?("test")
 
       load File.join(__dir__, "./fixtures/lib/strings.rb")
-      # This folder is not activated by FreezeTheLits
+      # This folder is not activated by Freezolite
       refute Strings.hello.frozen?
 
       load File.join(__dir__, "./fixtures/test/name_test.rb")