+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master
 
+- Change name from `freeze_the_lits` to `freezolite`. ([@palkan][])
+
+- Migrate to `require-hooks`. ([@palkan][])
+
 ## 0.1.0 (2023-06-15)
 
 - Initial. ([@palkan][])